+++ packages/crdt_lf/CHANGELOG.md
@@ -1,3 +1,8 @@
+## [Unreleased]
+
+### Added
+- Update operation for every handler
+
 ## [0.7.1] - 2025-06-26
 
 ### Changed